Add ToDos, Deadlines, and Events
Added the ToDo, Deadline, and Event subclasses, which inherit from the Task type, which each subclass having some slightly different functionality from each other. Used Regular Expressions for pattern matching in order to make sure that the command line input is clean, and the relevant fields such as the different time fields and task names could be picked out easily. Renamed the file and the overarching class to Barry.java and Barry respectively. I shifted the locations of some of the variables from the subclasses to the Task superclass, in order to preserve some OOP principles. I added accessor methods for all the private variables. I created a detailsAsString abstract method in the Task class, which is overriden by the subclasses to provide slightly different functionality. I created a help method, which returns a list of all available commands, which also indicates the correct format in which to use the command. I also error-proofed all commands, and used Exceptions in order to handle OutOfBounds Exceptions, when invalid array indexes were provided. I also added some very minor changes to the way that the chatbot responds, in order to make it more human-like. These include changes to the text output of the Barry chatbot, as well as adding pluralities in the event of more than one task. I also performed checks to ensure that a task was not already marked or unmarked, informing a user if a task was already marked/unmarked, and suggesting to use the unmark/mark command instead. I added a line of text output to the task adding commands that indicate the number of tasks in the list. I also removed the unused echo method. Each task adding command checks the command to see whether its in the correct format first, using Regular Expressions. If an incorrect format is detected, the chatbot will correctly identify the attempted command, based off the first few characters of the command, and provide help to the user accordingly in the form of a text output, directing the user in the correct direction. Finally, in the event that the user enters a command that does not match any known command, the chatbot now tells the user that it does not recognise the command, and suggests accessing the help menu once again.