Separate Parser into more specific classes
To allow Parser methods to return a Handler or Task object directly
without circular dependency. A marker interface is used as the Parser
objects do not share any methods.

Add use of enums for Task and Handler

Enum factory design pattern is used for Task and Handler object
creation for added clarity and type safety. Replaces use of switch
cases in Factory objects.