Add Sort Feature
Tasks are currently displayed on FIFO basis.

This results in a less than optimal user experience due to the
Tasks not being sorted by their deadline dates.

Sorting is done by placing tasks with earlier end dates at the top
of the list while tasks with later/no end dates (e.g ToDo Tasks) at
the bottom. This sorting is done by using a custom TaskComparator
class which does the sorting following the above requirements.

A TaskComparator is preffered so as to capitalise on the List.sort()
method.