Add Delete functionality
Added delete command which allows users to delete tasks in their list by number. Added similar error-handling functionaliy to that for the unmark command. Also removed redundant methods, and changed access modifiers of certain methods in the Task class to be protected, reflecting their status as protected methods. Also customised the delete message that is printed upon removing a task from the list.