Enums not added
I chose not to add an enums as adding enums would not improve the current functionality, efficiency, or the readability of the current program, just like how I did not need to define my own exceptions in the previous level as my code was able to handle all errors possible in the current condition of the program. I edited my help command to reflect the new delete command, and also edited the EXPECTED.txt file to reflect the new correct implementation of the help function.