Use assertions to document important assumptions
Add assertions for the default cases in switch statements in Parser and
Storage classes to indicate the assumption that one of the cases will
always be executed, i.e. user will only enter valid commands during
runtime.

An assertion for the load() method in Storage class is also added, as
the existing local file is assumed to have no empty strings.