Refactored and Reviewed Code Base
Parsing logic in Parser.java uses a series of case statements to
determine the correct response to a user's input. The function
itself looks very long and untidy.

I extracted out the lines and separated them into different
functions.

This is for readability and also to conceptually divide the
parsing into the types of commands that a user might give the
ChatBot.

I'm happy with the way the files are structured for now but in the
future we might want to organise the files into classes/objects
if the functionality expands and parsing becomes unfeasible or
too complicated.