Refactor code to improve code quality
Too many magic literals used in the previous commits. Instead, use named
constants to improve the readability of code.

Add more comments to specify what the methods are supposed to do, and
the rationale for the current implementations.

Use SLAP in switch statements in Parser class, while also applying
extract method to turn the fragment into a method whose name explains
the purpose of the method.