Add profile images
Rather than only having text in the dialogs which looks very plain,
added images in a resourse folder which will be displayed to visualise
the conversation between user and the program.

Updated the outputs on the Ui class.
Removed 1 assert line on the Storage class causing bugs.
Editted the main class to increase the size of the GUi.
Added images in a resourses/images folder.