Update build.gradle with new directory for application segment
The file, build.gradle, is outdated due to the newly implementation of
javaFx.

This change is required when creating a jar file.

The argument of mainClass is changed from "andelu.Andelu" to
"gui.Launcher".

Since the main entry of the program is changed due to javaFx, we
need to update the new main entry.