Add assert to the program
There is no assertion in the program to document important
assumptions.

This assertion helps to identify any potential error caused by
programmer.

Assertion has been added to each subclass of command.java and
storage.java.

The assumption made to these classes helps to ensure that the values
are correct or not null.