Improve code quality through refactoring
The main source of improvements was by moving the presentation logic
from `DukeApp` to `Ui`.

Previously, when responding to a user command, `DukeApp` would assume
that the responses would be preinted to `System.out` and redirect it
from there.

Now, `DukeApp` relies completely on `Ui` for displaying respones, and is
only responsible for executing commands.

Additionally, many minor changes were made in accordance with the Code
Style.