Add assertions to addTask and deleteTask methods. Additionally, added a private static final to remove magic constants.
To help to document the assumption that tasks to be added or deleted cannot be null/have invalid indexes.

To ensure that tasks are non-null and have valid index.

The assertion makes sure that if it is violated during runtime, an assertion error will be thrown, providing valuable information for debugging.