Refactor code to organise initialisation of task in TaskList
The previous code for initialisation of task was lengthy and hard to understand.

With these new changes, the initialisation of different types of task
are abstracted out for better readibility and flow. Logic wise, the code is the same.

This way, we are able to change the implementation of a specific types
initialisation without affecting others.