Add assert statements in Task and Duke
There is no verification that Duke is set up as intended. This may
result in a tasklist not being created, which is the point of the app
(to manage tasks). An assert statement is included in duke to ensure
that a tasklist is created; otherwise, it will prompt the programmer to
create a tasklist. Similarly, assert statements are added to Task to
verify that tasks are marked done (or not done) accurately.