Add undo extension
The following commands were added:
* REDO
* UNDO

which led to the update of the following classes to handle the new
commands:
*Command
*CommandProcessor

The package history was created with the following classes:
* History
* HistoryManager
* State

All Tasks were made to implement Cloneable and Serializable Interfaces
but the work has not been implemented fully. Deadline, Event still need
to be done properly.

Storage and Temp storage were updated to facilitate the undo and redo
commands but the actual functionality has not been done fully.