Add assertions in various points of the code
Potential bugs in the code might be present due to assumptions made in
the code.

If such bugs are present in an production environement, it will cause
the program to break or throw an uncaught exception.

Assertions are used to to detect violations in assumptions we have made
in the code. For example, we assert that a newly initialised task list
file contains "[]".

Implementing such assertions allows the runtime to verify that the
assumtions we made holds.

Let's
* Assert that the task list is NOT empty before printing it out to the user
* Assert that a new task list file contains the "[]" characters
* Assert that the user should not have an valid dateTime entry when
  processing its dateTime entry as a String