Fix bugs related to DoAfter command.
Many bugs were present in the previous commit, such as issues with
nesting, storage and retrieval from storage.

Fix nesting by simply removing reference to task when the task is
deleted from the taskList and updating methods so they do not
recursively call downwards when they encounter a DoAfter without
an associated task.

Fix storage and retrieval by allowing for population of generated
taskList during the retrieval phase and updating of storage syntax
for DoAfters.