Fix bad switch case statement in parser
Parse has a catch all default statement.

This is bad because default should be the default option
and not for the last option(s).

TODO, DEADLINE, EVENT cases have been consolidated to
call the addTask() method and default option is as intended.

Using default as the last option is an error prone way to
write code.

Please do not use default to catch remaining cases in the future.