Create jar file with gradle and standardise build.gradle file.
Following the textbook, I standardise the build.gradle file to utilise its functionality better, including creating fat jar files with shadowJar to capture external libraries like JavaFX. I also create a jar file that is executable.

This way, I make the building process easier and more standardised for future extension of this project. It is also easier for users or reviewers to run the jar file when it is pushed to remote repo.