Add assert statements at appropriate points to ensure certain assumptions made by the programmer are true.
This includes assertions in the TaskList and GrumbleBug files. They keep tabs on external classes and methods not implemented by this project such as StringBuilder and the .split String method, so we have no direct eyes on the details and we want to ensure we understand the endpoints' behavior correctly.