Add assertions to methods in AtsisBot and MainWindow classes
There are several methods that assume that certain conditions are true
and that the program will not crash if they are not. This is not a good.

This commit adds assertions to these methods to make sure that the
conditions are true and that the program will not crash if they are not.

This is being done by adding the assert keyword to the methods by
checking the conditions and throwing an AssertionError if the conditions
are not true.