Use streams to search for index of string in string array
A for loop is used to search for a string in an array.

This method is less readable because the happy path and intent to find
the index of the first occurring string is not clear.

Use streams to search for the first occurrence of the string instead.

This way, the intent is clear and the default case of returning -1 is
obvious.