Tidy code and fix bug
Overall comment for the last few commits.
Code quality is lacking.

The code needs to be maintainable and easily understandable.

Changed names to be easier to understand, improved
switch cases, improved inline comments and removed
magic numbers.

This will help people to understand the code better and
reduce the chance of bugs and misunderstandings.

Because I was combing through for code quality, I realised
a big bug for the equality of two TaskLists.