Add assertions into Parser and TaskList
The initialize method in Parser assumes that the input string to produce
the type of tasks is going to be just one single string. The mark,
unmark, and delete method of the TaskList assumes that the index input
are going to be greater than one.

There are just implicit assumption and are not checked while running
the program. As such, it is uncertain whether if these assumptions
hold true.

Using asserts statements make these assumptions explicit and they are
checked while the program is compiled.

Let's add some assert statements regarding this into the Parser and
TaskList class.