Add Graphical User Interface to app
The chatbot uses the terminal to communicate.

This might not be practical with non-tech
savvy users, and the terminal isn't great
to look/use sometimes.

Add a GUI using JavaFX to provide for users
to visually look at their input and output.
Add JavaFX FXML files and their respective
controllers to help abstract out elements
in the GUI for easier readability and
maintainability.

This makes it more intuitive that it is a
Chatbot, and maybe easier for certain users.