Currently, there is no usage of assert.
It requires assert at Main and DialogBox so that users and the programme understands the assumptions that I made.

I used assert to ensure that all the assumption made are catch and the user using it will understand.

This is done so that my chatbot does not go into runtime error / null pointer error.

This is in addition to the A-Assertions commit made earlier