Add comments to improve code readability
There are not comments for method in FileParser and InputParser.

It is difficult to understand why strings are split multiple times.

Let's explain why strings a split in such a way - according to the
format of the input string.

The comments will help developers understand and edit the code if there
are changes to the app commands.