Include assertion statements
Update GUI welcome message
Bee GUI shows welcome message when user inputs 'hello' command
This causes a blank GUI to be shown when user launches the chatbot.
This can be confusing to the user as they do not know what they are
suppose to do.

The initialise method in MainWindow now calls the dialogContainer to
create a new dialogBox with the bee image and the welcome
message to be printed out when the program is ran.