Refactor Storage
File path for the chatbot's data is hard-coded.

This is less safe since it is a magic literal. Changing the file path would require many repeated changes.

Moreover, the file path should be a constant.

Let us extract the `filePath` variable to be a static variable associated with `Storage` class.