Improve code quality
If-else blocks are used heavily throughout the loadTasks method in the
Storage class. The same statement that checks the content of each line
in the text file is used repeatedly at every else-if block.

Let's clean up the code by replacing the if-else statements with a
switch statement so that the code looks more tidy and readable.

This reduces repetition as well. Making the code look more readable and clean allows other people to
better understand it and improves the maintainability of the code.