Fix "mark" and "unmark" commands for Steven
Steven has commands to mark (or unmark) a given task as "complete"
or "incomplete" respectively.

There is a bug which respectively caused mark and unmark to mark
given task as "incomplete" or "complete", and also prevents
tasks as being marked or unmarked when incomplete and complete
respectively

This is unintended behavior.

Bugfix should invert the above behavior

Let's
* allow marking to be done on incomplete tasks, and set them to
complete
* unmarking to be done on complete tasks, and set them to
incomplete