Add assertions to test important assumptions
There are seveal methods that assumes that the argument passed in is
valid. In general usage this is true, however there may be cases where
that is forgetten and an invalid value may be passed into the function.
This can cause the program to perform erroneously to continue.

As a fail safe for such a catastrophic event, let's add several
assertion to indicate to us what had gone wrong.