Fix commit issues
This fixes issues relating to commits

I originally created a commit with assertions. However, I
merged the PR early, so I undid the changes in the
commit. However, when I fixed it and went to create
the PR at the correct time, the changes I had initially
merger were not reflected in the PR, but were on the
branch.

This commit fixes that issue by editing each of the
original assertions slightly

This is so that git recognises it as a new change and
hence registers them in the new PR.