Implement date clash feature
The current version of the chatbot does not check for date clashes for
the event task.

This is bad for the user because the user does not know that an
existing event clashes with the event the user wants to add.

Let's update the necessary classes in order to implement the check for
date clashes for event tasks.