Add assertions
Add assertions to multiple methods in different classes to test for
assumptions.

There are no assertions to determine if there is an error in programming
logic.

Adding assertions allows us to verify the correctness of the programming
logic. We will be able to identify the error early and prevent the
effect from cascading further.

Let's add assertions to multiple methods in different classes to test
our assumptions.