Add assertions in key places
Currently, some parts of the code operate on assumptions.
For example getting a task from an ArrayList and assuming it is not null.

Adding assertions for unlikely but possible events provide more safety.

This will help us in catching errors in code.
More so if we change some of these methods that are key to the assumptions.