Remove deprecated methods in TaskList Class
Currently, the process of saving and storing the task list has been
abstracted to the Storage class. However, the methods for both can still
be found in the TaskList class.

This is a problem as it violates the Single Responsibility Principle and
creates confusion for people reviewing the code.

Lets address this by deleting both methods from the TaskList class
directly.