Update build.gradle
build.gradle use the new main class

The old main class which was on Duke.java is changed to Launcher.java.

The main class is needed to create the new JAR version. Therefore, without changing it
the JAR file cannot run. By changing it, the JAR file can run fine.