Refactor parseCommand method in Parser class
Currently, the parseCommand method is very lengthy as it handles the
logic for each possible case.

This should be avoided as it makes the method too long, in violation of
the style guide.

Lets change the parser class so that it has separate methods to handle
each type of command word, which parseCommand invokes.