Change names of methods for Parser Class
There are identical names for methods for the Parser class and the
TaskList class.

Changing the names for the Parser class and TaskList class can improve
code quality, as the method names now provide a clearer idea of what the
method does to the respective classes.

Let's changes the names of addTask methods of Parser to
addTaskToTaskList and addTask respectively.

The reader will be able to
understand that Parser is modifying the instance of TaskList, while
the TaskList is getting modified through the usage of the AddTask
method.