Refactor parser class
Parser class: Extract method to check arguments length

The Parser class has several methods that checks if the
arguments supplied by the user satisfy a minimum length.

This is duplicated code that can be extracted into a single
method. This allows for additional validation should there
be any that is required in future.