Change parseCommand to return String
I am aware that this is not an ideal way of doing things, since parseCommand is not supposed to interact with UI directly or return anything. However, given the scope of the project and the amount of time I want to spend here, I believe this amendment is fine for now. It is not good practice.