Add deadline tasks sorting
All tasks are currently appended to the end of the
task list by default, with no way for the user to sort them.

Deadline tasks possess an inherent trait, a deadline that
can be used to chronologically sort them.

The new command "sort-deadlines" sorts tasks of type Deadline
in place.

Since it is debatable whether or not it is logically appropriate
to sort todos and events, we avoid them completely and simply
re-arrange the deadline tasks in-place.