Add detection for duplicate tasks
There is no detection for duplicate tasks entered by the user.

This may cause confusion to the users as they use the list function
and see duplicate tasks.

Let's add checks to detect duplicate tasks within the TaskList.

The checks will check on the behalf of the users for duplicate tasks.