Add assert to Storage.
In the method parseText(), the if statements only checks for the
task type. It is assumed that the correct number of parameters has
been passed to the function.

Assert has been added to check that the correct number of parameters
are given.

To prevent any incorrect inputs given to the method.