Implement B-ViewSchedules
Current implementation allows users to view their schedule on a
specific date using the "on" command.

This features provides a way for users to view tasks on a specific date
in the form of a schedule.

Let's,
* modify the OnCommand class and rename it to ViewCommand,
refactoring the current code
* modify the Parser class such that the "view" command displays the
scheduled tasks on the given date instead of the "on" command
* modify the JUnit test cases in the ParserTest class to account for the
modified "view" command and the refactored ViewCommand class