Fix bug in file creation
The pervious version does not have a functionality for creating the
directory where the save file is to be stored in. This would cause
errors whenever such a condition occurs.

This commit fixes the error and creates a directory if there is no
directory to currently store the save file.