Complete undo extension
Completed undo extension by integrating a start state that is set at
start up. The following classes were modified:
* History
* HistoryManager
* State
There was a bug where objects in the state were still pointing to the
same values, so cloning objects was not a deepcopy. This was fixed by
using the clone method in TempStorage::restore as well

Removed Cloneable interface and methods in Task and Todo.