Integrate Checkstyle to project
Integrate Checkstyle to the project with Gradle and refactor code that contained violations.

The goal is to have a tool to automatically check the code for any coding standard violations.

Note that following config files and how it affects the checks:
config/checkstyle/checkstyle.xml: Rules to follow.
config/checkstyle/suppressions.xml: Rules to suppress under in which files.