Use assert in Ui to ensure important assumptions
There are several methods in Ui where tasks will be added and deleted.
As a result, the size of the tasklist will go up and down accordingly.

The usage of assert will ensure that the size of these will adjust
according to what task has been presented, whether it is a delete or
add.