Different levels of abstraction for methods in DialogBox and MainWindow class
Should be changed so as to abide with SLAP and improve readability of code.

Abstracted the setting of box alignment code into a new method in DialogBox class.
Code for displaying dialog boxes also abstracted into a new method in MainWindow class.

This ensures that the level of abstractions are the same, abiding with SLAP.