Sort task list
Task list is unsorted and the order depends on the time the task was
added. User may find it difficult to locate tasks if list is long.

Let's add a command for sort that sorts tasks by their task type and
deadline. If task type is deadline, they will be located at the start of
the list and sorted by their deadline starting from the earliest
deadline to latest.

This allows user to locate tasks easily, and can be reminded of the
tasks with the earliest deadline so that they can complete them on time.